rename de `/router/` a `/routers/`
