+++ backend/src/router.ts
@@ -6,19 +6,19 @@ import { PrismaAuthRepository } from "./repositories/auth.js";
 import { PrismaSuscripcionRepository } from "./repositories/suscripciones.js";
 import { AuthServiceImpl } from "./services/auth.js";
 import { SuscripcionServiceImpl } from "./services/suscripciones.js";
-import { authRouter } from "./router/auth.js";
-import { suscripcionesRouter } from "./router/suscripciones.js";
+import { authRouter } from "./routers/auth.js";
+import { suscripcionesRouter } from "./routers/suscripciones.js";
 import { SuscripcionHandler } from "./handlers/suscripciones.js";
 import { AdministradorHandler } from "./handlers/administrador.js";
 import { PrismaAdministradorRepository } from "./repositories/administrador.js";
 import { AdministradorServiceImpl } from "./services/administrador.js";
-import { administradoresRouter } from "./router/administrador.js";
+import { administradoresRouter } from "./routers/administrador.js";
 import { EstablecimientoHandler } from "./handlers/establecimientos.js";
 import { PrismaEstablecimientoRepository } from "./repositories/establecimientos.js";
 import { EstablecimientoServiceImpl } from "./services/establecimientos.js";
-import { establecimientosRouter } from "./router/establecimientos.js";
+import { establecimientosRouter } from "./routers/establecimientos.js";
 import { AuthMiddleware } from "./middlewares/auth.js";
-import { canchasRouter } from "./router/canchas.js";
+import { canchasRouter } from "./routers/canchas.js";
 import { PrismaCanchaRepository } from "./repositories/canchas.js";
 import { CanchaServiceimpl } from "./services/canchas.js";
 import { CanchaHandler } from "./handlers/canchas.js";
