+++ backend/src/routers/administrador.ts
